requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
httpx>=0.25.0
aiohttp>=3.9.0
anyio>=4.0.0

# AI / LLM
anthropic>=0.7.0
openai>=1.3.0
tiktoken>=0.5.0

# Inference (vLLM - GPU only, commented for CPU dev)
# vllm>=0.2.0
# torch>=2.1.0

# Storage
redis[hiredis]>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0

# Monitoring / Observability
prometheus-client>=0.18.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
structlog>=23.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
ujson>=5.8.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Integrations
google-api-python-client>=2.100.0
requests>=2.31.0
